feat: relocate mapper methods for list_files() method
